refactor: split Screen classes into separate files
Screen 클래스를 개별 파일로 분리하고 main.py를 진입점으로 축소:
- screens/start.py: StartScreen
- screens/server_config.py: ServerConfigScreen
- screens/client_config.py: ClientConfigScreen
- screens/chat.py: ChatScreen (TCPConnection 콜백 통합)
- app.py: TCPSocketApp
- main.py: ~6줄 진입점으로 축소
